tools/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
GeoAlchemy2==0.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
lxml==5.3.0
numpy==2.1.3
packaging==24.2
psycopg2-binary==2.9.10
pyproj==3.7.0
python-dotenv==1.0.1
shapely==2.0.6
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
